Fix compiler warnings on x32
tv_nsec is __syscall_slong_t == long long there, so we need a cast.

* libdb/db_lookup.c (dbprintf): Cast tv_nsec to long for %ld format.
* libdb/db_store.c (make_content): Likewise.
* src/check_mandirs.c (testmandirs, update_db): Likewise.
* src/man.c (maybe_update_file): Likewise.

+++ libdb/db_lookup.c
@@ -116,7 +116,7 @@ void dbprintf (const struct mandata *info)
 	       "whatis:    %s\n\n",
 	       dash_if_unset (info->name),
 	       info->ext, info->sec, info->comp,
-	       info->id, (long) info->mtime.tv_sec, info->mtime.tv_nsec,
+	       info->id, (long) info->mtime.tv_sec, (long) info->mtime.tv_nsec,
 	       info->pointer, info->filter, info->whatis);
 }
